Implement PR #90 Changes
