+++ Testing/VelaptorTests/Hardware/SystemMonitorTests.cs
@@ -108,11 +108,6 @@ namespace VelaptorTests.Hardware
                 Height = 200,
             };
 
-            var myMonitor = new SystemMonitor(null)
-            {
-                Width = 123,
-            };
-
             // Act
             var actual = monitor.Center;
