+++ Velaptor/Input/KeyboardState.cs
@@ -406,9 +406,9 @@ namespace Velaptor.Input
 
             var keys = Enum.GetValues(typeof(KeyCode)).Cast<KeyCode>().ToArray();
 
-            foreach (var k in keys)
+            foreach (var key in keys)
             {
-                this.keyStates.Add(k, false);
+                this.keyStates.Add(key, false);
             }
         }
     }
