+++ Velaptor/OpenGL/Enums.cs
@@ -2,10 +2,30 @@
 // Copyright (c) KinsonDigital. All rights reserved.
 // </copyright>
 
-using Velaptor.NativeInterop.OpenGL;
-
 namespace Velaptor.OpenGL
 {
+    using Velaptor.NativeInterop.OpenGL;
+
+    /// <summary>
+    /// Represents the type of GPU buffer.
+    /// </summary>
+    public enum BufferType
+    {
+        /// <summary>
+        ///  A vertex buffer object.
+        /// </summary>
+        VertexBufferObject,
+
+        /// <summary>
+        /// An index array object.
+        /// </summary>
+        /// <remarks>
+        ///     Also known as an EBO (Element Buffer Object) or
+        ///     IBO (Index Buffer Object).
+        /// </remarks>
+        IndexArrayObject,
+    }
+
     /// <summary>
     /// Specifies the program parameter.
     /// <para>
@@ -38,8 +58,8 @@ namespace Velaptor.OpenGL
     /// Specifies the target to which the buffer object is bound for the buffer data.
     /// <para>Used in the following OpenGL function calls.</para>
     /// <list type="bullet">
-    ///     <item><see cref="IGLInvoker.BufferData(GLBufferTarget, uint, nint, GLBufferUsageHint)"/></item>
-    ///     <item><see cref="IGLInvoker.BufferSubData{T}(GLBufferTarget, nint, nuint, ref T)"/></item>
+    ///     <item><see cref="IGLInvoker.BufferData(Velaptor.OpenGL.GLBufferTarget,float[],Velaptor.OpenGL.GLBufferUsageHint)"/></item>
+    ///     <item><see cref="IGLInvoker.BufferSubData"/></item>
     /// </list>
     /// </summary>
     internal enum GLBufferTarget
@@ -58,7 +78,7 @@ namespace Velaptor.OpenGL
     /// <summary>
     /// Specifies the expected usage pattern of the data store.
     /// <para>
-    ///     Used for the OpenGL function <see cref="IGLInvoker.BufferData(GLBufferTarget, uint, nint, GLBufferUsageHint)"/>.
+    ///     Used for the OpenGL function <see cref="IGLInvoker.BufferData(Velaptor.OpenGL.GLBufferTarget,float[],Velaptor.OpenGL.GLBufferUsageHint)"/>.
     /// </para>
     /// </summary>
     internal enum GLBufferUsageHint
@@ -111,7 +131,7 @@ namespace Velaptor.OpenGL
     /// <summary>
     /// Specifies the number of color components in the texture.
     /// <para>
-    ///     Used for the OpenGL function <see cref="IGLInvoker.TexImage2D{T}(GLTextureTarget, int, GLInternalFormat, uint, uint, int, GLPixelFormat, GLPixelType, void*)"/>.
+    ///     Used for the OpenGL function <see cref="IGLInvoker.TexImage2D{T}"/>.
     /// </para>
     /// </summary>
     internal enum GLInternalFormat
@@ -125,7 +145,7 @@ namespace Velaptor.OpenGL
     /// <summary>
     /// Specifies the format of the pixel data.
     /// <para>
-    ///     Used for the OpenGL function <see cref="IGLInvoker.TexImage2D{T}(GLTextureTarget, int, GLInternalFormat, uint, uint, int, GLPixelFormat, GLPixelType, void*)"/>.
+    ///     Used for the OpenGL function <see cref="IGLInvoker.TexImage2D{T}"/>.
     /// </para>
     /// </summary>
     internal enum GLPixelFormat
@@ -139,7 +159,7 @@ namespace Velaptor.OpenGL
     /// <summary>
     /// Specifies the format of the pixel data.
     /// <para>
-    ///     Used for the OpenGL function <see cref="IGLInvoker.TexImage2D{T}(GLTextureTarget, int, GLInternalFormat, uint, uint, int, GLPixelFormat, GLPixelType, void*)"/>.
+    ///     Used for the OpenGL function <see cref="IGLInvoker.TexImage2D{T}"/>.
     /// </para>
     /// </summary>
     internal enum GLPixelType
@@ -172,15 +192,10 @@ namespace Velaptor.OpenGL
         /// OpenGL enum name GL_DEBUG_OUTPUT.
         /// </summary>
         DebugOutput = 37600,
-
-        /// <summary>
-        /// OpenGL enum name GL_LINE_SMOOTH.
-        /// </summary>
-        LineSmooth = 2848,
     }
 
     /// <summary>
-    /// Specifices how the red, green, blue, and alpha source blending factors are computed.
+    /// Specifies how the red, green, blue, and alpha source blending factors are computed.
     /// <para>
     ///     Used for the OpenGL function <see cref="IGLInvoker.BlendFunc(GLBlendingFactor, GLBlendingFactor)"/>.
     /// </para>
@@ -201,21 +216,6 @@ namespace Velaptor.OpenGL
         /// OpenGL enum name GL_ONE_MINUS_SRC_ALPHA.
         /// </summary>
         OneMinusSrcAlpha = 771,
-
-        /// <summary>
-        /// OpenGL enum name GL_ONE_MINUS_DST_COLOR.
-        /// </summary>
-        OneMinusDestColor = 773,
-
-        /// <summary>
-        /// OpenGL enum name GL_CONSTANT_COLOR.
-        /// </summary>
-        ConstantColor = 32769,
-
-        /// <summary>
-        /// OpenGL enum name GL_ONE_MINUS_CONSTANT_COLOR.
-        /// </summary>
-        OneMinusConstantColor = 32770,
     }
 
     /// <summary>
@@ -235,156 +235,6 @@ namespace Velaptor.OpenGL
         /// OpenGL enum name GL_TEXTURE1.
         /// </summary>
         Texture1,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE2.
-        /// </summary>
-        Texture2,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE3.
-        /// </summary>
-        Texture3,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE.
-        /// </summary>
-        Texture4,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE5.
-        /// </summary>
-        Texture5,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE6.
-        /// </summary>
-        Texture6,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE7.
-        /// </summary>
-        Texture7,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE8.
-        /// </summary>
-        Texture8,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE9.
-        /// </summary>
-        Texture9,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE10.
-        /// </summary>
-        Texture10,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE11.
-        /// </summary>
-        Texture11,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE12.
-        /// </summary>
-        Texture12,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE13.
-        /// </summary>
-        Texture13,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE14.
-        /// </summary>
-        Texture14,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE15.
-        /// </summary>
-        Texture15,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE16.
-        /// </summary>
-        Texture16,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE17.
-        /// </summary>
-        Texture17,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE18.
-        /// </summary>
-        Texture18,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE19.
-        /// </summary>
-        Texture19,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE20.
-        /// </summary>
-        Texture20,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE21.
-        /// </summary>
-        Texture21,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE22.
-        /// </summary>
-        Texture22,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE23.
-        /// </summary>
-        Texture23,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE24.
-        /// </summary>
-        Texture24,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE25.
-        /// </summary>
-        Texture25,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE26.
-        /// </summary>
-        Texture26,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE27.
-        /// </summary>
-        Texture27,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE28.
-        /// </summary>
-        Texture28,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE29.
-        /// </summary>
-        Texture29,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE30.
-        /// </summary>
-        Texture30,
-
-        /// <summary>
-        /// OpenGL enum name GL_TEXTURE31.
-        /// </summary>
-        Texture31,
     }
 
     /// <summary>
@@ -392,8 +242,8 @@ namespace Velaptor.OpenGL
     /// <para>Used in the following OpenGL function calls.</para>
     /// <list type="bullet">
     ///     <item><see cref="IGLInvoker.BindTexture(GLTextureTarget, uint)"/></item>
-    ///     <item><see cref="IGLInvoker.TexParameter(GLTextureTarget, GLTextureParameterName, int)"/></item>
-    ///     <item><see cref="IGLInvoker.TexImage2D{T}(GLTextureTarget, int, GLInternalFormat, uint, uint, int, GLPixelFormat, GLPixelType, void*)"/></item>
+    ///     <item><see cref="IGLInvoker.TexParameter(Velaptor.OpenGL.GLTextureTarget,Velaptor.OpenGL.GLTextureParameterName,Velaptor.OpenGL.GLTextureWrapMode)"/></item>
+    ///     <item><see cref="IGLInvoker.TexImage2D{T}"/></item>
     /// </list>
     /// </summary>
     internal enum GLTextureTarget
@@ -407,7 +257,7 @@ namespace Velaptor.OpenGL
     /// <summary>
     /// Specifies what kind of primitives to render.
     /// <para>
-    ///     Unsed in the <see cref="IGLInvoker.DrawElements(GLPrimitiveType, uint, GLDrawElementsType, System.IntPtr)"/>.
+    ///     Unused in the <see cref="IGLInvoker.DrawElements(GLPrimitiveType, uint, GLDrawElementsType, System.IntPtr)"/>.
     /// </para>
     /// </summary>
     internal enum GLPrimitiveType
@@ -508,235 +358,34 @@ namespace Velaptor.OpenGL
     internal enum GLDebugSource
     {
         /// <summary>
-        /// OpenGL enum name GL_DEBUG_SOURCE_API.
-        /// </summary>
-        DebugSourceApi = 33350,
-
-        /// <summary>
-        /// OpenGL enum name GL_DEBUG_SOURCE_WINDOW_SYSTEM.
-        /// </summary>
-        DebugSourceWindowSystem = 33351,
-
-        /// <summary>
-        /// OpenGL enum name GL_DEBUG_SOURCE_SHADER_COMPILER.
-        /// </summary>
-        DebugSourceShaderCompiler = 33352,
-
-        /// <summary>
-        /// OpenGL enum name GL_DEBUG_SOURCE_THIRD_PARTY.
-        /// </summary>
-        DebugSourceThirdParty = 33353,
-
-        /// <summary>
         /// OpenGL enum name GL_DEBUG_SOURCE_APPLICATION.
         /// </summary>
         DebugSourceApplication = 33354,
-
-        /// <summary>
-        /// OpenGL enum name GL_DEBUG_SOURCE_OTHER.
-        /// </summary>
-        DebugSourceOther = 33355,
-    }
-
-    /// <summary>
-    /// Used in GL.Arb.DebugMessageControl, GL.Arb.DebugMessageInsert and 6 other functions.
-    /// </summary>
-    internal enum GLDebugType
-    {
-        /// <summary>
-        /// OpenGL enum name GL_DEBUG_TYPE_ERROR.
-        /// </summary>
-        DebugTypeError = 33356,
-
-        /// <summary>
-        /// OpenGL enum name GL_DEBUG_TYPE_DEPRECATED_BEHAVIOR.
-        /// </summary>
-        DebugTypeDeprecatedBehavior = 33357,
-
-        /// <summary>
-        /// OpenGL enum name GL_DEBUG_TYPE_UNDEFINED_BEHAVIOR.
-        /// </summary>
-        DebugTypeUndefinedBehavior = 33358,
-
-        /// <summary>
-        /// OpenGL enum name GL_DEBUG_TYPE_PORTABILITY.
-        /// </summary>
-        DebugTypePortability = 33359,
-
-        /// <summary>
-        /// OpenGL enum name GL_DEBUG_TYPE_PERFORMANCE.
-        /// </summary>
-        DebugTypePerformance = 33360,
-
-        /// <summary>
-        /// OpenGL enum name GL_DEBUG_TYPE_OTHER.
-        /// </summary>
-        DebugTypeOther = 33361,
-
-        /// <summary>
-        /// OpenGL enum name GL_DEBUG_TYPE_MARKER.
-        /// </summary>
-        DebugTypeMarker = 33384,
-
-        /// <summary>
-        /// OpenGL enum name GL_DEBUG_TYPE_PUSH_GROUP.
-        /// </summary>
-        DebugTypePushGroup = 33385,
-
-        /// <summary>
-        /// OpenGL enum name GL_DEBUG_TYPE_POP_GROUP.
-        /// </summary>
-        DebugTypePopGroup = 33386,
-    }
-
-    /// <summary>
-    /// Used in GL.Arb.DebugMessageControl, GL.Arb.DebugMessageInsert and 6 other functions.
-    /// </summary>
-    internal enum GLDebugSeverity
-    {
-        /// <summary>
-        /// OpenGL enum name GL_DEBUG_SEVERITY_NOTIFICATION
-        /// </summary>
-        DebugSeverityNotification = 33387,
-
-        /// <summary>
-        /// OpenGL enum name GL_DEBUG_SEVERITY_HIGH.
-        /// </summary>
-        DebugSeverityHigh = 37190,
-
-        /// <summary>
-        /// OpenGL enum name GL_DEBUG_SEVERITY_MEDIUM.
-        /// </summary>
-        DebugSeverityMedium = 37191,
-
-        /// <summary>
-        /// OpenGL enum name GL_DEBUG_SEVERITY_LOW.
-        /// </summary>
-        DebugSeverityLow = 37192,
-    }
-
-    // TODO: Move this into the GLFW folder in an Enums.cs file and to the same namespace
-    /// <summary>
-    /// Error codes, used in the GLFW error callback.
-    /// </summary>
-    internal enum GLFWErrorCode
-    {
-        /// <summary>
-        /// Everything is running as intended. Yay!
-        /// </summary>
-        NoError = 0,
-
-        /// <summary>
-        /// Called a function before calling IGlfw.Init. Initialize GLFW and then try again.
-        /// </summary>
-        NotInitialized = 65537,
-
-        /// <summary>
-        /// No OpenGL/OpenGL ES context on this thread.
-        /// </summary>
-        NoContext = 65538,
-
-        /// <summary>
-        /// Used an invalid enum value on a function.
-        /// </summary>
-        /// <remarks>
-        /// This should hopefully never happen in the bindings, due to the
-        /// added type safety of C# enums VS. GLFW's native #defines.
-        /// </remarks>
-        InvalidEnum = 65539,
-
-        /// <summary>
-        /// Called a function with an invalid argument.
-        /// </summary>
-        /// <remarks>
-        ///     This can happen if you request an OpenGL version that doesn't exist, like 2.7.
-        ///     If you request a version of OpenGL that exists, but isn't supported by this graphics
-        ///     card, it will return VersionUnavailable instead.
-        /// </remarks>
-        InvalidValue = 65540,
-
-        /// <summary>
-        /// A memory allocation failed on GLFW's end.
-        /// </summary>
-        /// <remarks>
-        ///     Report this to the GLFW issue tracker if encountered.
-        /// </remarks>
-        OutOfMemory = 65541,
-
-        /// <summary>
-        /// The requested API is not available on the system.
-        /// </summary>
-        ApiUnavailable = 65542,
-
-        /// <summary>
-        /// The requested OpenGL version is not available on the system.
-        /// </summary>
-        VersionUnavailable = 65543,
-
-        /// <summary>
-        /// A platform-specific error occurred that doesn't fit into any more specific category.
-        /// </summary>
-        /// <remarks>
-        ///     Report this to the GLFW issue tracker if encountered.
-        /// </remarks>
-        PlatformError = 65544,
-
-        /// <summary>
-        /// The requested format is unavailable.
-        /// </summary>
-        /// <remarks>
-        ///     If emitted during window creation, the requested pixel format isn't available.
-        ///     If emitted when using the clipboard, the contents of the clipboard couldn't be
-        ///     converted to the requested format.
-        /// </remarks>
-        FormatUnavailable = 65545,
-
-        /// <summary>
-        /// There is no OpenGL/OpenGL ES context attached to this window.
-        /// </summary>
-        NoWindowContext = 65546,
     }
 
     /// <summary>
     /// Bitwise OR of masks that indicate the buffers to be cleared using the follow masks.
-    /// <list type="bullet"
+    /// <list type="bullet">
     ///     <item><see cref="ColorBufferBit"/></item>
-    ///     <item><see cref="DepthBufferBit"/></item>
-    ///     <item><see cref="StencilBufferBit"/></item>
     /// </list>
     /// </summary>
     internal enum GLClearBufferMask
     {
         /// <summary>
-        /// Indicates the depth buffer.
-        /// </summary>
-        DepthBufferBit = 0x100,
-
-        /// <summary>
-        /// Indicates the stencil buffer.
-        /// </summary>
-        StencilBufferBit = 0x400,
-
-        /// <summary>
         /// Indicates the buffers currently enabled for color writing.
         /// </summary>
         ColorBufferBit = 0x4000,
-
-        /// <summary>
-        /// Unknown.
-        /// </summary>
-        CoverageBufferBitNV = 0x8000,
     }
 
     /// <summary>
-    /// Defineds behavior of how to render textures that are outside of
+    /// Defines behavior of how to render textures that are outside of
     /// the NDC (Normalized Device Coordinates) of 0 to 1.
     /// </summary>
     internal enum GLTextureWrapMode
     {
         /// <summary>
         /// Clamps the coordinates between 0 and 1.  The result is that higher coordinates
-        /// become clamped tot he edge, resulting in a streched edge pattern.
+        /// become clamped tot he edge, resulting in a stretched edge pattern.
         /// </summary>
         ClampToEdge = 33071,
     }
@@ -746,11 +395,6 @@ namespace Velaptor.OpenGL
     /// </summary>
     internal enum GLTextureMagFilter
     {
-        /// <summary>
-        /// Returns the value of the texture element that is nearest (in Manhattan distance) to the specified texture coordinates.
-        /// </summary>
-        Nearest = 9728,
-
         /// <summary>
         /// Returns the weighted average of the four texture elements that are closest to the specified
         /// texture coordinates. These can include items wrapped or repeated from other parts of a texture,
@@ -766,36 +410,11 @@ namespace Velaptor.OpenGL
     internal enum GLTextureMinFilter
     {
         /// <summary>
-        /// Returns the value of the texture element that is nearest (in Manhattan distance) to the specified texture coordinates.
-        /// </summary>
-        Nearest = 9728,
-
-        /// <summary>
         /// Returns the weighted average of the four texture elements that are closest to the specified
         /// texture coordinates. These can include items wrapped or repeated from other parts of a texture,
         /// depending on the values of <see cref="GLTextureParameterName.TextureWrapS"/> and
         /// <see cref="GLTextureParameterName.TextureWrapT"/>, and on the exact mapping.
         /// </summary>
         Linear = 9729,
-    }
-
-    /// <summary>
-    /// Represents the type of GPU buffer.
-    /// </summary>
-    public enum BufferType
-    {
-        /// <summary>
-        ///  A vertex buffer object.
-        /// </summary>
-        VertexBufferObject,
-
-        /// <summary>
-        /// An index array object.
-        /// </summary>
-        /// <remarks>
-        ///     Also known as an EBO (Element Buffer Object) or
-        ///     IBO (Index Buffer Object).
-        /// </remarks>
-        IndexArrayObject,
     }
 }
