+++ Velaptor/OpenGL/GPUBufferNameAttribute.cs
@@ -6,14 +6,21 @@ namespace Velaptor.OpenGL
 {
     using System;
 
+    /// <summary>
+    /// Represents the name of a buffer.
+    /// </summary>
     [AttributeUsage(AttributeTargets.Class)]
     internal class GPUBufferNameAttribute : Attribute
     {
-        public GPUBufferNameAttribute(string name)
-        {
-            Name = name;
-        }
+        /// <summary>
+        /// Initializes a new instance of the <see cref="GPUBufferNameAttribute"/> class.
+        /// </summary>
+        /// <param name="name">The name to give a buffer.</param>
+        public GPUBufferNameAttribute(string name) => Name = name;
 
+        /// <summary>
+        /// Gets the name of a buffer.
+        /// </summary>
         public string Name { get; }
     }
 }
